Add project for Maxim targets
This project is added for MAX32630 target. The source code should remain same
for other targets as long as the register addresses and bits remain same.
When exporting the flash blob in build/c_blob.c file to other project
(e.g. ARMmbed/DAPLink), the values of structure 'device_cfg_t' must be changed
to the corresponding addresses of the target post export.
Be sure to change value of 'verify' field in program_target_t structure to 0. A
garbage value is loaded due to an error.